fix: improve error handling in WebSocket connection closure and client management
