+++ internal/handlers/websocket_test.go
@@ -38,8 +38,12 @@ func TestWebSocketUpgrade(t *testing.T) {
 	if err != nil {
 		t.Fatalf("WebSocket connection failed: %v", err)
 	}
-	defer conn.Close()
-	defer resp.Body.Close()
+	if err := conn.Close(); err != nil {
+		t.Fatal(err)
+	}
+	if err := resp.Body.Close(); err != nil {
+		t.Fatal(err)
+	}
 
 	// Verify client is registered
 	if h.ClientsCount() != 1 {
