+++ internal/hub/hub.go
@@ -2,10 +2,7 @@
 package hub
 
 import (
-	"log"
 	"sync"
-
-	"github.com/gorilla/websocket"
 )
 
 // Hub represents a hub for managing WebSocket connections.
@@ -62,48 +59,6 @@ func (h *Hub) Run() {
 				}
 			}
 			h.mu.RUnlock() // Unlock after reading
-		}
-	}
-}
-
-// NewClient creates a new Client instance with the provided hub and WebSocket connection.
-func NewClient(hub *Hub, conn *websocket.Conn) *Client {
-	return &Client{
-		hub:  hub,
-		conn: conn,
-		send: make(chan []byte, 256),
-	}
-}
-
-// ReadPump handles reading messages from the WebSocket connection and broadcasting them to the hub.
-func (c *Client) ReadPump() {
-	defer func() {
-		c.hub.Unregister <- c
-		if err := c.conn.Close(); err != nil {
-			log.Printf("can't close the connection: %v", err)
-		}
-	}()
-	for {
-		_, message, err := c.conn.ReadMessage()
-		if err != nil {
-			break
-		}
-		c.hub.broadcast <- message
-	}
-}
-
-// WritePump handles writing messages from the client's send channel to the WebSocket connection.
-// WritePump handles writing messages from the client's send channel to the WebSocket connection.
-func (c *Client) WritePump() {
-	defer func() {
-		c.hub.Unregister <- c
-		if err := c.conn.Close(); err != nil {
-			log.Printf("can't close the connection: %v", err)
-		}
-	}()
-	for message := range c.send {
-		if err := c.conn.WriteMessage(websocket.TextMessage, message); err != nil {
-			break
 		}
 	}
 }
